Address PR #74 review: trim changelog, document VM URL forms, split types
- CHANGELOG: collapse the verification-relationship entry to 3-4 sentences
  (full rationale stays in PRD §10.5) and remove the duplicate `### Added`
  block that listed it twice.
- IVerificationRelationshipResolver: document the three accepted forms of
  `verificationMethodDidUrl` (absolute DID URL, fragment-relative `#k1`,
  bare `k1`) and that ordinal comparison does not strip query/path.
- Split AuthorizationDecision and VerificationRelationshipAuthorizationResult
  into their own files for discoverability, matching the one-type-per-file
  convention elsewhere in NetDid.Core.
- Add two tests pinning down the documented normalization on the query
  side: fragment-only and bare-id `verificationMethodDidUrl` both match a
  controller-qualified entry.
