+++ src/NetDid.Core/Resolution/IVerificationRelationshipResolver.cs
@@ -3,48 +3,6 @@ using NetDid.Core.Model;
 namespace NetDid.Core.Resolution;
 
 /// <summary>
-/// The decision returned by <see cref="IVerificationRelationshipResolver"/>. The tri-state shape
-/// keeps "infrastructure failed" distinct from "controller's document said no", so a verifier can
-/// fail closed and log the underlying cause instead of silently treating both as denial.
-/// </summary>
-public enum AuthorizationDecision
-{
-    Authorized,
-    NotAuthorized,
-    ControllerNotResolvable
-}
-
-/// <summary>
-/// Outcome of a verification-relationship authorization check. <see cref="ResolutionError"/> and
-/// <see cref="Message"/> are populated only when <see cref="Decision"/> is
-/// <see cref="AuthorizationDecision.ControllerNotResolvable"/>.
-/// </summary>
-public sealed record VerificationRelationshipAuthorizationResult
-{
-    public required AuthorizationDecision Decision { get; init; }
-
-    /// <summary>The underlying resolution error code (e.g. "notFound", "invalidDid", "methodNotSupported").</summary>
-    public string? ResolutionError { get; init; }
-
-    /// <summary>Human-readable diagnostic message; never logs key material.</summary>
-    public string? Message { get; init; }
-
-    public static VerificationRelationshipAuthorizationResult Authorized() =>
-        new() { Decision = AuthorizationDecision.Authorized };
-
-    public static VerificationRelationshipAuthorizationResult NotAuthorized() =>
-        new() { Decision = AuthorizationDecision.NotAuthorized };
-
-    public static VerificationRelationshipAuthorizationResult NotResolvable(string error, string message) =>
-        new()
-        {
-            Decision = AuthorizationDecision.ControllerNotResolvable,
-            ResolutionError = error,
-            Message = message
-        };
-}
-
-/// <summary>
 /// Answers, against the <em>controller's</em> DID document, whether a verification method is
 /// authorized for a given W3C verification relationship.
 /// Used by zcap-dotnet for ZCAP-LD controller-authorization checks (proof-purpose →
@@ -64,6 +22,16 @@ public interface IVerificationRelationshipResolver
     /// the document's id — equals <paramref name="verificationMethodDidUrl"/>. Matches both
     /// referenced and embedded entries.
     /// </summary>
+    /// <param name="controllerDid">Absolute DID of the controller whose document is consulted
+    /// (e.g. <c>"did:web:example.com"</c>). Must be non-empty.</param>
+    /// <param name="verificationMethodDidUrl">Identifier of the verification method to check.
+    /// Accepts three forms, all normalized against the resolved controller id before comparison:
+    /// an absolute DID URL (<c>"did:web:alice.example#key-1"</c>), a fragment-relative
+    /// reference (<c>"#key-1"</c> → <c>"{controllerDid}#key-1"</c>), or a bare id
+    /// (<c>"key-1"</c> → <c>"{controllerDid}#key-1"</c>). Comparison is ordinal and does NOT
+    /// strip query parameters or path segments.</param>
+    /// <param name="relationship">The W3C verification relationship to consult.</param>
+    /// <param name="ct">Cancellation token.</param>
     /// <exception cref="ArgumentException">Thrown when <paramref name="controllerDid"/> or
     /// <paramref name="verificationMethodDidUrl"/> is null or whitespace.</exception>
     Task<VerificationRelationshipAuthorizationResult> IsAuthorizedForRelationshipAsync(
